refactor(orm): split aggregate error codes, RLS guards and index definitions into leaf modules
PERF-001 prep. query.ts throws AGGREGATE_ERROR/COUNT_ERROR in ~200 places and
lifecycle.ts reads index definitions; both must stay reachable without the
3k-line aggregate runtime. An unfiltered count() is served by the native Convex
syscall, so its RLS guard moves too.
